[BUGFIX] Ensure MDC usage when configured
If the MDC usage is enabled within the driver configuration the MDC
should always be enabled by default and only be disabled if a proper
configuration can be retrieved. Within the backend context there are
many requests that cannot retrieve any Site information. Thus the MDC
setting cannot be evaluated and is turned on by default.